ParkSense-Backend/requirements.txt
Flask==2.3.0
Flask-RESTX==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pytest==7.3.1
pytest-cov==4.1.0
requests==2.31.0
python-dateutil==2.8.2
APScheduler==3.10.4
